Add all function in class
