+++ src/modules/api.js
@@ -4,26 +4,34 @@ const likeUrl = `https://us-central1-involvement-api.cloudfunctions.net/capstone
 
 // API for post Comment
 
-const postComment = async (index, username, comment) => {
-  await fetch(commentUrl, {
-    method: 'POST',
-    body: JSON.stringify({
-      item_id: index,
-      username,
-      comment,
-    }),
-    headers: { 'content-Type': 'application/json; charset=UTF-8' },
-  });
-};
+class Movies {
+  getMovies = async () => {
+    try {
+      const response = await fetch('https://api.tvmaze.com/shows');
+      const data = await response.json();
+      return data;
+    } catch (error) {
+      return error;
+    }
+  }
 
-// API for get comment
+  postComment = async (index, username, comment) => {
+    await fetch(commentUrl, {
+      method: 'POST',
+      body: JSON.stringify({
+        item_id: index,
+        username,
+        comment,
+      }),
+      headers: { 'content-Type': 'application/json; charset=UTF-8' },
+    });
+  }
 
-const getComment = async (index) => {
-  const response = await fetch(`${commentUrl}?item_id=${index}`);
-  const user = await response.json();
-  return user;
-};
+  getComment = async (index) => {
+    const response = await fetch(`${commentUrl}?item_id=${index}`);
+    const user = await response.json();
+    return user;
+  };
+}
 
-export {
-  postComment, getComment,
-};
+export default Movies;
